Refactor code s.t. workarounds are in respective packages rather than a separate workarounds package
