+++ constants/constants.go
@@ -15,7 +15,6 @@ const (
 	CNIConfigDir                          = "/etc/cni"
 	CNIStateDir                           = "/var/lib/cni"
 	SystemdDir                            = "/etc/systemd/system"
-	ConfigDir                             = "conf"
 	FlannelVersion                        = "v0.10.0"
 	DefaultPodNetwork                     = "10.244.0.0/16"
 	DefaultDNSIP                          = "10.96.0.10"
